MAINT: refactor/clean up some GenomeData types
