+++ q2_types/feature_data_mag/_format.py
@@ -9,7 +9,6 @@
 import re
 
 from q2_types.feature_data import DNAFASTAFormat
-from q2_types.genome_data._format import OrthologFileFmt
 from qiime2.plugin import model
 
 from ..plugin_setup import plugin
@@ -62,30 +61,3 @@ class MAGSequencesDirFmt(model.DirectoryFormat):
 
 
 plugin.register_formats(MAGSequencesDirFmt)
-
-
-class OrthologAnnotationDirFmt(model.DirectoryFormat):
-    pathspec = r'.+\.annotations'
-    annotations = model.FileCollection(pathspec, format=OrthologFileFmt)
-
-    @annotations.set_path_maker
-    def annotations_path_maker(self, file_name):
-        return file_name.split(sep="_")[0]
-
-    def annotation_dict(self, relative=False) -> dict:
-        ids = {}
-        for path in self.path.iterdir():
-            if re.compile(self.pathspec).match(path.name):
-                _id = re.sub('.emapper$', '', path.stem)
-                absolute_path = path.absolute()
-                if relative:
-                    ids[_id] = str(
-                        absolute_path.relative_to(self.path.absolute())
-                    )
-                else:
-                    ids[_id] = str(absolute_path)
-
-        return dict(sorted(ids.items()))
-
-
-plugin.register_formats(OrthologAnnotationDirFmt)
