+++ q2_types/feature_data_mag/_transformer.py
@@ -5,22 +5,17 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
-from io import StringIO
-
-import uuid
-
 import collections.abc
 import glob
 import os.path
 
-import qiime2
 from itertools import repeat
 
 import pandas as pd
 import skbio
 from q2_types.feature_data._transformer import _fastaformats_to_series
 
-from . import MAGSequencesDirFmt, OrthologAnnotationDirFmt
+from . import MAGSequencesDirFmt
 from ..plugin_setup import plugin
 
 CONSTRUCTORS = {
@@ -94,84 +89,3 @@ def _5(data: MAGIterator) -> MAGSequencesDirFmt:
         with open(fp, 'a') as fin:
             skbio.io.write(seq, format='fasta', into=fin)
     return result
-
-
-def _is_valid_uuid4(uuid_string: str):
-    """
-    Check if a given string is a valid UUID version 4.
-
-    This function checks if the provided string is a valid UUID version 4.
-    The only purpose of doing that here is to identify whether provided
-    string was a MAG ID (UUID4) or a sample ID. For that reason, we don't
-    print any statements or raise any exceptions.
-
-    Parameters:
-    uuid_string (str): The string to check for UUID version 4 validity.
-
-    Returns:
-    bool: True if the string is a valid UUID version 4, False otherwise.
-    """
-    try:
-        uuid_obj = uuid.UUID(uuid_string, version=4)
-    except ValueError:
-        return False
-
-    return str(uuid_obj) == uuid_string
-
-
-def _reshuffle_columns(df: pd.DataFrame):
-    if 'MAG' in df.columns:
-        col = 'MAG'
-    elif 'Sample' in df.columns:
-        col = 'Sample'
-    else:
-        return df
-
-    cols = df.columns.tolist()
-    cols.remove(col)
-    cols.insert(0, col)
-    df = df[cols]
-    return df
-
-
-def _annotations_to_dataframe(
-        data: OrthologAnnotationDirFmt
-) -> pd.DataFrame:
-    annotations = data.annotation_dict()
-    dfs = []
-    for _id, path in annotations.items():
-        # we need to ignore the ## comments at
-        # the beginning and end of the file
-        with open(path, 'r') as f:
-            lines = [line for line in f if not line.startswith('##')]
-
-        df = pd.read_csv(StringIO('\n'.join(lines)), sep='\t', index_col=0)
-        if _is_valid_uuid4(_id):
-            df['MAG'] = _id
-        else:
-            df['Sample'] = _id
-
-        dfs.append(df)
-
-    df = pd.concat(dfs)
-
-    # to satisfy QIIME2's particular requirements
-    df.reset_index(drop=False, inplace=True)
-    df.index = df.index.astype(str)
-    df.index.rename('id', inplace=True)
-
-    # reshuffle columns for nicer display
-    df = _reshuffle_columns(df)
-
-    return df
-
-
-@plugin.register_transformer
-def _7(data: OrthologAnnotationDirFmt) -> pd.DataFrame:
-    return _annotations_to_dataframe(data)
-
-
-@plugin.register_transformer
-def _8(data: OrthologAnnotationDirFmt) -> qiime2.Metadata:
-    annotations = _annotations_to_dataframe(data)
-    return qiime2.Metadata(annotations)
