+++ q2_types/feature_data_mag/tests/test_format.py
@@ -11,11 +11,8 @@ import shutil
 import unittest
 
 from qiime2.plugin.testing import TestPluginBase
-from qiime2.plugin import ValidationError
 
-from q2_types.feature_data_mag._format import (
-        MAGSequencesDirFmt, OrthologAnnotationDirFmt,
-        )
+from q2_types.feature_data_mag._format import MAGSequencesDirFmt
 
 
 class TestFormats(TestPluginBase):
@@ -57,39 +54,6 @@ class TestFormats(TestPluginBase):
                 '3b7d53fb-5b60-46c6-8819-aeda065b12e9.fasta',
             '6232c7e1-8ed7-47c8-9bdb-b94706a26931':
                 '6232c7e1-8ed7-47c8-9bdb-b94706a26931.fasta',
-        }
-        self.assertDictEqual(obs, exp)
-
-    def test_ortholog_annotation_dir_fmt_passing(self):
-        dirpath = self.get_data_path('good_ortholog_annotation')
-        fmt_obj = OrthologAnnotationDirFmt(dirpath, mode='r')
-        fmt_obj.validate()
-
-    def test_ortholog_annotation_dir_fmt_fails_extra_file(self):
-        dirpath = self.get_data_path('ortholog_annotation_extra')
-        fmt_obj = OrthologAnnotationDirFmt(dirpath, mode='r')
-
-        with self.assertRaisesRegex(ValidationError, "Unrecognized file"):
-            fmt_obj.validate()
-
-    def test_ortholog_annotations_annot_dict(self):
-        annotations = OrthologAnnotationDirFmt(
-            self.get_data_path('ortholog_annotation_samples'), mode='r'
-        )
-
-        obs = annotations.annotation_dict()
-        exp = {
-            'test_output1':
-                str(annotations.path / 'test_output1.emapper.annotations'),
-            'test_output2':
-                str(annotations.path / 'test_output2.emapper.annotations')
-        }
-        self.assertDictEqual(obs, exp)
-
-        obs = annotations.annotation_dict(relative=True)
-        exp = {
-            'test_output1': 'test_output1.emapper.annotations',
-            'test_output2': 'test_output2.emapper.annotations'
         }
         self.assertDictEqual(obs, exp)
