+++ q2_types/genome_data/_type.py
@@ -7,22 +7,29 @@
 # ----------------------------------------------------------------------------
 
 from qiime2.plugin import SemanticType
-from q2_types.feature_data import BLAST6
 
 from . import (
     GenesDirectoryFormat, ProteinsDirectoryFormat, LociDirectoryFormat,
-    SeedOrthologDirFmt,
+    SeedOrthologDirFmt, GenomeSequencesDirectoryFormat,
+    OrthologAnnotationDirFmt,
 )
 from ..plugin_setup import plugin
+from ..sample_data import SampleData
 
-GenomeData = SemanticType('GenomeData', field_names='type',
-                          field_members={'type': BLAST6})
+GenomeData = SemanticType('GenomeData', field_names='type')
 Genes = SemanticType('Genes', variant_of=GenomeData.field['type'])
 Proteins = SemanticType('Proteins', variant_of=GenomeData.field['type'])
 Loci = SemanticType('Loci', variant_of=GenomeData.field['type'])
-Ortholog = SemanticType('Ortholog', variant_of=GenomeData.field['type'])
+Orthologs = SemanticType('Orthologs',
+                         variant_of=[GenomeData.field['type'],
+                                     SampleData.field['type']])
+NOG = SemanticType('NOG', variant_of=[GenomeData.field['type'],
+                                      SampleData.field['type']])
+DNASequence = SemanticType('DNASequence', variant_of=GenomeData.field['type'])
 
-plugin.register_semantic_types(GenomeData, Genes, Proteins, Loci)
+plugin.register_semantic_types(
+    GenomeData, Genes, Proteins, Loci, DNASequence, Orthologs, NOG
+)
 
 plugin.register_semantic_type_to_format(
     GenomeData[Genes],
@@ -40,6 +47,26 @@ plugin.register_semantic_type_to_format(
 )
 
 plugin.register_semantic_type_to_format(
-    GenomeData[BLAST6],
+    GenomeData[Orthologs],
     artifact_format=SeedOrthologDirFmt
+)
+
+plugin.register_semantic_type_to_format(
+    GenomeData[DNASequence],
+    artifact_format=GenomeSequencesDirectoryFormat
+)
+
+plugin.register_semantic_type_to_format(
+    GenomeData[NOG],
+    artifact_format=OrthologAnnotationDirFmt
+)
+
+plugin.register_semantic_type_to_format(
+    SampleData[Orthologs],
+    artifact_format=SeedOrthologDirFmt
+)
+
+plugin.register_artifact_class(
+    SampleData[NOG],
+    directory_format=OrthologAnnotationDirFmt
 )
