+++ q2_types/genome_data/tests/data/good_ortholog_annotation/test_output.emapper.annotations
