+++ q2_types/genome_data/tests/data/ortholog_annotation_mags/d33dbcbe-eedd-4a19-8b93-e76a15ce107f.emapper.annotations
