+++ q2_types/genome_data/tests/data/ortholog_annotation_mags/e6408340-b308-4325-83a1-ea406f3e19ad.emapper.annotations
