+++ q2_types/genome_data/tests/test_format.py
@@ -14,6 +14,7 @@ from qiime2.plugin.testing import TestPluginBase
 from .._format import (
     GenesDirectoryFormat, ProteinsDirectoryFormat, GFF3Format,
     LociDirectoryFormat, SeedOrthologDirFmt, OrthologFileFmt,
+    OrthologAnnotationDirFmt,
 )
 
 
@@ -138,6 +139,39 @@ class TestFormats(TestPluginBase):
         with self.assertRaisesRegex(
                 ValidationError, 'The phase on line 10 was 8.'):
             GFF3Format(filepath, mode='r').validate()
+
+    def test_ortholog_annotation_dir_fmt_passing(self):
+        dirpath = self.get_data_path('good_ortholog_annotation')
+        fmt_obj = OrthologAnnotationDirFmt(dirpath, mode='r')
+        fmt_obj.validate()
+
+    def test_ortholog_annotation_dir_fmt_fails_extra_file(self):
+        dirpath = self.get_data_path('ortholog_annotation_extra')
+        fmt_obj = OrthologAnnotationDirFmt(dirpath, mode='r')
+
+        with self.assertRaisesRegex(ValidationError, "Unrecognized file"):
+            fmt_obj.validate()
+
+    def test_ortholog_annotations_annot_dict(self):
+        annotations = OrthologAnnotationDirFmt(
+            self.get_data_path('ortholog_annotation_samples'), mode='r'
+        )
+
+        obs = annotations.annotation_dict()
+        exp = {
+            'test_output1':
+                str(annotations.path / 'test_output1.emapper.annotations'),
+            'test_output2':
+                str(annotations.path / 'test_output2.emapper.annotations')
+        }
+        self.assertDictEqual(obs, exp)
+
+        obs = annotations.annotation_dict(relative=True)
+        exp = {
+            'test_output1': 'test_output1.emapper.annotations',
+            'test_output2': 'test_output2.emapper.annotations'
+        }
+        self.assertDictEqual(obs, exp)
 
 
 if __name__ == '__main__':
